refactor, fix bugs
